requirements.txt
attrs==23.1.0
bounded-pool-executor==0.0.3
circuitgraph==0.2.1
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
distro==1.8.0
docopt==0.6.2
edalize==0.5.1
fastjsonschema==2.17.1
fonttools==4.39.4
fusesoc==2.2.1
Jinja2==3.1.2
jsonschema==4.17.3
kiwisolver==1.4.4
lark==1.1.5
libcomps
liberty-parser==0.0.16
MarkupSafe==2.1.3
matplotlib==3.7.1
mpmath==1.3.0
networkx==3.1
numpy==1.24.3
okonomiyaki==1.3.2
packaging==23.1
pandas==2.0.2
Pillow==9.5.0
pqdm==0.2.0
pykwalify==1.8.0
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
scipy==1.11.2
simplesat==0.8.2
six==1.16.0
sqlitedict==2.1.0
sympy==1.6.2
tqdm==4.65.0
typing_extensions==4.6.3
tzdata==2023.3
vcdvcd
zipfile2==0.0.12
scipy
sqlitedict
lmdb
west==1.1.0
